Connectors moved to deprecated
