+++ packages/sdk/typedoc.json
@@ -4,5 +4,6 @@
   "entryPoints": ["src/index.ts"],
   "categoryOrder": ["Bridge deposit", "Event handling", "Fee", "Token iteractions", "Helpers"],
   "exclude": ["**/*+(.test|.e2e).ts"],
+  "readme": "../../README.md",
   "includeVersion": true
 }
